+++ client/client.go
@@ -16,7 +16,7 @@ import (
 	"github.com/fxamacker/cbor/v2"
 	"github.com/gofiber/fiber/v3/log"
 
-	"github.com/gofiber/utils/v2"
+	utils "github.com/gofiber/utils/v2"
 
 	"github.com/valyala/fasthttp"
 	"github.com/valyala/fasthttp/fasthttpproxy"