+++ middleware/logger/utils.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gofiber/fiber/v3"
 	fiberlog "github.com/gofiber/fiber/v3/log"
-	"github.com/gofiber/utils/v2"
+	utils "github.com/gofiber/utils/v2"
 )
 
 func methodColor(method string, colors fiber.Colors) string {