+++ path.go
@@ -15,7 +15,7 @@ import (
 	"time"
 	"unicode"
 
-	"github.com/gofiber/utils/v2"
+	utils "github.com/gofiber/utils/v2"
 	"github.com/google/uuid"
 )